Modularize whole project into dedicated go packages
